+++ web/src/pages/project/AnalysisRunnerView/Filter.ts
@@ -1,0 +1,4 @@
+export type Filter = {
+    category: string
+    value: string
+}
